Move RegexManager to a dedicated file
